fix: move SearchResult from tests/contracts to domain.py (#5)
- Add SearchResult dataclass to services/marketplace_service/domain.py
- Update __init__.py to import SearchResult from domain (remove try/except)
- Update tests/contracts/marketplace to re-export from domain.py
- Change packages field type to List[Dict[str, Any]] to avoid test contract dependency

Closes #5
